Check in gradle wrapper files
This project uses gradle. We are going to check in the gradle wrapper files (two script files and a small gradle-wrapper.jar and gradle-wrapper.properties).  This way other users do not need to download and install the full gradle distribution to start using gradle with this project.

Some may find it a bit odd that we would check in a jar file into Git, but this the recommended way: c.f.  https://docs.gradle.org/current/dsl/org.gradle.api.tasks.wrapper.Wrapper.html